Move unit system to util (#2763)
