+++ homeassistant/bootstrap.py
@@ -12,8 +12,6 @@ from typing import Any, Optional, Dict
 
 import voluptuous as vol
 
-from homeassistant.helpers.typing import HomeAssistantType
-
 import homeassistant.components as core_components
 from homeassistant.components import group, persistent_notification
 import homeassistant.config as conf_util
@@ -218,7 +216,7 @@ def prepare_setup_platform(hass: core.HomeAssistant, config, domain: str,
 
 # pylint: disable=too-many-branches, too-many-statements, too-many-arguments
 def from_config_dict(config: Dict[str, Any],
-                     hass: Optional[HomeAssistantType]=None,
+                     hass: Optional[core.HomeAssistant]=None,
                      config_dir: Optional[str]=None,
                      enable_log: bool=True,
                      verbose: bool=False,
