+++ homeassistant/helpers/service.py
@@ -1,21 +1,19 @@
 """Service calling related helpers."""
 import functools
 import logging
-
 # pylint: disable=unused-import
 from typing import Optional  # NOQA
 
 import voluptuous as vol
 
-from homeassistant.helpers.typing import HomeAssistantType  # NOQA
-
 from homeassistant.const import ATTR_ENTITY_ID
+from homeassistant.core import HomeAssistant  # NOQA
 from homeassistant.exceptions import TemplateError
 from homeassistant.helpers import template
 from homeassistant.loader import get_component
 import homeassistant.helpers.config_validation as cv
 
-HASS = None  # type: Optional[HomeAssistantType]
+HASS = None  # type: Optional[HomeAssistant]
 
 CONF_SERVICE = 'service'
 CONF_SERVICE_TEMPLATE = 'service_template'
